refactor: create secondary endpoints
