+++ src/commands/package.json
@@ -1,0 +1,11 @@
+{
+  "ngPackage": {
+    "lib": {
+      "entryFile": "./public_api.ts",
+      "umdModuleIds": {
+        "prosemirror-schema-list": "prosemirrorSchemaList",
+        "prosemirror-commands": "prosemirrorCommands"
+      }
+    }
+  }
+}
