+++ src/commands/toggleList.ts
@@ -2,7 +2,7 @@ import { EditorState, Transaction } from 'prosemirror-state';
 import { NodeType } from 'prosemirror-model';
 import { wrapInList, liftListItem } from 'prosemirror-schema-list';
 
-import { isNodeActive } from '../helpers';
+import { isNodeActive } from 'ngx-editor/helpers';
 
 export const toggleList = (type: NodeType, itemType: NodeType) => {
   return (state: EditorState, dispatch: (tr: Transaction) => void) => {
