+++ src/commands/toggleWrap.ts
@@ -2,7 +2,7 @@ import { EditorState, Transaction } from 'prosemirror-state';
 import { NodeType } from 'prosemirror-model';
 import { wrapIn, lift } from 'prosemirror-commands';
 
-import isNodeActive from '../helpers/isNodeActive';
+import { isNodeActive } from 'ngx-editor/helpers';
 
 export const toggleWrap = (type: NodeType) => {
   return (state: EditorState, dispatch: (tr: Transaction) => void) => {
