+++ src/helpers/isListItem.ts
