+++ src/helpers/isNodeActive.ts
