+++ src/helpers/package.json
@@ -1,0 +1,7 @@
+{
+  "ngPackage": {
+    "lib": {
+      "entryFile": "./public_api.ts"
+    }
+  }
+}
