+++ src/helpers/public_api.ts
@@ -2,3 +2,4 @@ export * from './isMarkActive';
 export * from './isNodeActive';
 export * from './isListItem';
 export * from './getSelectionMarks';
+export * from './bubblePosition';
