+++ src/lib/types.ts
@@ -1,24 +1,6 @@
-import { Plugin, EditorState, Transaction } from 'prosemirror-state';
+import { Plugin } from 'prosemirror-state';
 import { Node as ProsemirrorNode, Schema } from 'prosemirror-model';
 import { EditorView, Decoration, NodeView } from 'prosemirror-view';
-
-type TCR = { dom: HTMLElement, update: (state: EditorState) => void };
-
-export type ToolbarDropdown = { heading?: string[] };
-export type ToolbarCustomMenuItem = (editorView: EditorView) => TCR;
-export type ToolbarDropdownGroupKeys = keyof ToolbarDropdown;
-export type ToolbarDropdownGroupValues = ToolbarDropdown[ToolbarDropdownGroupKeys];
-export type ToolbarItem = string | ToolbarDropdown | ToolbarCustomMenuItem;
-export type Toolbar = Array<ToolbarItem[]> | null;
-
-export interface MenuItemViewSpec {
-  classNames?: string[];
-  innerHTML?: string;
-  textContent?: string;
-  attrs?: { [key: string]: string };
-  activeClass: string;
-  disabledClass: string;
-}
 
 export interface NodeViews {
   [name: string]: (
@@ -33,23 +15,4 @@ export interface NgxEditorConfig {
   schema?: Schema;
   plugins?: Plugin[];
   nodeViews?: NodeViews;
-}
-
-export type MenuLabels = { [key: string]: string };
-export interface MenuOptions {
-  toolbar?: Toolbar;
-  labels?: MenuLabels;
-  schema?: Schema;
-}
-
-export type Command = (schema: EditorState, dispatch: (tr: Transaction) => void) => boolean;
-
-export interface DropdownViewRender {
-  dom: HTMLElement;
-  updates: Array<(state: EditorState) => void>;
-}
-
-export interface MenuItemViewRender {
-  dom: HTMLElement;
-  update: (state: EditorState) => void;
 }
