+++ src/plugins/package.json
@@ -1,0 +1,12 @@
+{
+  "ngPackage": {
+    "lib": {
+      "entryFile": "./public_api.ts",
+      "umdModuleIds": {
+        "prosemirror-state": "prosemirrorState",
+        "prosemirror-view": "prosemirrorView",
+        "prosemirror-commands": "prosemirrorCommands"
+      }
+    }
+  }
+}
