+++ src/plugins/public_api.ts
@@ -1,3 +1,5 @@
 export { default as placeholder } from './placeholder';
 export { default as menu } from './menu';
 export { default as link } from './link';
+
+export * from './types';
