+++ src/plugins/utils/icons/bold.ts
