+++ src/plugins/utils/icons/italic.ts
