+++ src/plugins/utils/icons/quote.ts
