+++ wiki/content/configuartion.md
@@ -1,7 +1,8 @@
 # Configuration
 
 ```ts
-import { menu, placeholder, schema } from 'ngx-editor';
+import { schema } from 'ngx-editor';
+import { menu, placeholder } from 'ngx-editor/plugins';
 
 NgxEditorModule.forRoot({
   schema, // optional scheama, see https://sibiraj.dev/ngx-editor/additional-documentation/schema.html
